requirements.txt
Pillow>=12.0.0,<13
pywebview>=6.1
scikit-image>=0.25
